**** 1st stage **** 1. Start Screen (0-1 ... 0-3)
2. Map&Menu (1, 1-2, 1-3)
 * About screen instead Help (show simple dialog) - Copyright OsmAnd 2014
 * 1-3 Show current street name (call Alexey function)
 * When location is unknown and you click on my location(show a hint "location is not known yet" like a toast)
 * Q. Slide right doesn't work as closing
 * Шрифт пунктов меню сейчас не тот, сейчас видимо Helvetica, Нужен Avenir.

5. Favorites (3-1-1 ... 3-1-5)
 * Add to favorites (crash)
 * Some favorites without color (screenshot)
 * No arrow direction to favorites
 * 3-1-2... 3-1-5 not testable (can't access these screens),
 * Mini map could be screenshot

7. Settings (4-1-1 ... 4-1-2)
 * Russian language? Should be english
 * Hide zoom buttons

+++ Sources/Models/OAFavoriteItem.h
@@ -12,7 +12,7 @@
 @interface OAFavoriteItem : NSObject
 
 @property std::shared_ptr<OsmAnd::IFavoriteLocation> favorite;
-@property float direction;
+@property CGFloat direction;
 @property NSString* distance;
